requirements.txt
# requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
tenacity
apscheduler
pyyaml
python-multipart
python-jose

# AWS
boto3
botocore
boto3-stubs

# Azure
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-web
azure-mgmt-sql
azure-mgmt-containerservice
azure-mgmt-containerinstance
azure-mgmt-keyvault
azure-mgmt-monitor
azure-mgmt-authorization
azure-mgmt-resource

# GCP
google-cloud-compute
google-cloud-storage
google-auth
google-auth-oauthlib
google-cloud-container
google-cloud-functions
google-cloud-resource-manager
google-cloud-asset

# Google IAM
google-cloud-iam
google-cloud-run
google-cloud-build
google-cloud-compute
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
google-cloud-container
google-cloud-functions
google-cloud-spanner
google-cloud-dataproc
google-cloud-firestore
google-cloud-redis
google-cloud-bigtable
google-cloud-dns
google-cloud-logging
google-cloud-monitoring
google-cloud-filestore
google-cloud-iap
google-cloud-resource-manager
google-cloud-billing
google-cloud-tasks
google-cloud-scheduler
google-cloud-dataflow-client
# Development
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy
types-requests
# typing-extensions must be >=4.14.1 to satisfy pydantic>=2.12.0
typing-extensions

# Recommended Google client libraries (ensure typing support)
google-api-python-client
google-api-core
google-auth-httplib2

# Type stubs
types-python-jose
types-pyyaml

# Optional: install specific google-cloud-* packages if needed (uncomment to pin)
# google-cloud-pubsub==3.14.0
# google-cloud-dns==0.36.0
# google-cloud-billing==1.20.0
# google-cloud-redis==2.11.0
# google-cloud-firestore==2.10.0